Fix UnicodeDecodeError in cache query
When the environment in UTF-8 the getattr() call fails occasionally with
a UnicodeDecodeError. Instead of requiring our Clear systems to be on a
specific locale or to call DNF with a specific locale, just skip this
read. The rest of the function already gracefully handles empty values
by continuing so val is just set to None so dnf does not fail with a
traceback.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>