Add workaround for proxy environment variable issue
The https_proxy environment variable was not being recognized by dnf
(via librepo), which was somehow triggered by setting the librepo proxy
config to the empty string. Since librepo uses libcurl, and libcurl
normally honors this proxy environment variable, the root cause of the
problem must lie elsewhere, but this workaround is simple enough to use
for now.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>